chore: reorganization & improvements
